feat(cli): update dev script to compile in watch mode
