+++ cli/build.ts
@@ -1,45 +1,23 @@
 import fs from 'fs'
 import path from 'path'
 
-import { COMPONENTS_DIRECTORY, MOULD_DIRECTORY } from './constants'
+import compile from './compile'
+import {
+    COMPONENTS,
+    COMPONENTS_DIRECTORY,
+    MOULD_DIRECTORY,
+    SCHEMA,
+} from './constants'
 
-import { transform } from '../compile/transform'
-
-const originalDirectory: string = process.cwd()
+const originalDirectory = process.cwd()
 
 const appPath = path.join(__dirname, '..')
-const componentsPath = path.join(appPath, COMPONENTS_DIRECTORY, 'index.js')
+const componentsPath = path.join(appPath, COMPONENTS_DIRECTORY, COMPONENTS)
 const mouldPath = path.join(originalDirectory, MOULD_DIRECTORY)
-const schemaPath = path.join(mouldPath, '.mould')
+const schemaPath = path.join(mouldPath, SCHEMA)
 
 if (fs.existsSync(schemaPath)) {
-    const time = process.hrtime()
-
-    fs.readFile(schemaPath, 'utf8', (err, schema) => {
-        if (err) {
-            console.error('Failed to read Mould Schema\n' + err)
-            process.exit(1)
-        }
-
-        fs.writeFile(
-            componentsPath,
-            transform(JSON.parse(schema)),
-            'utf8',
-            (err) => {
-                if (err) {
-                    console.error('Failed to write Mould Components\n' + err)
-                    process.exit(1)
-                }
-
-                const [s, ns] = process.hrtime(time)
-                console.log(
-                    `Compiled Mould Components successfully in ${s}s ${
-                        ns / 1e6
-                    }ms`
-                )
-            }
-        )
-    })
+    compile(schemaPath, componentsPath)
 } else if (fs.existsSync(mouldPath)) {
     console.warn(
         `You don't have Mould Schema at ${mouldPath}\n\n` +
